# docs/openapi.yaml
openapi: 3.0.3
info:
  title: STAAD Report Extractor API
  version: 1.0.0
  description: |
    FastAPI service for uploading a STAAD `.std` model, queueing extraction work,
    running the selected structural-analysis backend, and polling for results.

    ANALYSIS_BACKEND=inhouse is the default; openstaad retains the licensed
    Windows oracle. In-house health does not require Bentley licensing. In-house
    station distances use metres and source units are N/METER. The analysis block
    adds backend, solver_version, validation_status and numerical diagnostics.
    This is a validation build; see docs/VALIDATION_RESULTS.md for differences.

    Runtime architecture:

    `Client → API → Queue → Worker → In-house solver or OpenSTAAD → Result`

    Extraction flows (`POST /jobs` `extraction_flow`):

    - `standard` (default, Sliding) — global absolute envelope under `result.properties`
    - `fully_unitized` — same global envelope plus per-profile blocks and `result.profiles`
    - `casement` — same standard-shaped global envelope plus a top-level signed
      `result.casement` object (`interlock`, `central_meeting`, `fixed_mullion`,
      `horizontal`, `outer`)

    Authentication is not required. `POST /jobs` uses `multipart/form-data`.
  contact:
    name: STAAD Report Extractor
servers:
  - url: http://localhost:8000
    description: Local development (default Uvicorn)
...
paths:
...
  /jobs:
    post:
      tags: [Jobs]
      operationId: createJob
      summary: Submit a STAAD extraction job
      description: |
        Accepts a `.std` upload, stages it under the jobs directory, and enqueues
        work for the background worker. Returns immediately with HTTP 202; the
        request path does **not** launch or attach to STAAD.

        Poll `GET /jobs/{job_id}` for progress. Call `GET /jobs/{job_id}/result`
        only when `result_ready` is `true`.

        Set `extraction_flow=casement` for Casement signed profile envelopes. The
        optional `generation_request` JSON may include `casement_profiles` to
        override geometry classification.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/JobCreateRequest"
            encoding:
              std_file:
                contentType: application/octet-stream
              generation_request:
                contentType: application/json
      responses:
        "202":
          description: Job accepted and queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobCreateResponse"
              example:
                job_id: 3f8c2a1e-9b4d-4e6a-8c1f-2a7b9d0e5f11
                status: queued
                message: Job accepted.
        "413":
          description: Uploaded file exceeds `STAAD_API_MAX_UPLOAD_BYTES` (default 50 MB)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetailEnvelope"
              example:
                detail:
                  error_code: file_too_large
                  message: Upload exceeds maximum allowed size.
                  retryable: false
        "422":
          description: Validation failed (file type, timeout, retries, flow, or path)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ErrorDetailEnvelope"
                  - $ref: "#/components/schemas/ValidationErrorResponse"
              examples:
                invalidInputFile:
                  summary: Not a .std file
                  value:
                    detail:
                      error_code: invalid_input_file
                      message: Only .std file upload is supported.
                      retryable: false
                invalidTimeout:
                  summary: timeout_seconds <= 0
                  value:
                    detail:
                      error_code: invalid_timeout
                      message: timeout_seconds must be > 0
                      retryable: false
                invalidMaxRetries:
                  summary: max_retries < 0
                  value:
                    detail:
                      error_code: invalid_max_retries
                      message: max_retries must be >= 0
                      retryable: false
                invalidRetryDelay:
                  summary: retry_delay_seconds < 1
                  value:
                    detail:
                      error_code: invalid_retry_delay
                      message: retry_delay_seconds must be >= 1
                      retryable: false
                invalidExtractionFlow:
                  summary: Unknown extraction_flow
                  value:
                    detail:
                      error_code: invalid_extraction_flow
                      message: extraction_flow must be 'standard', 'fully_unitized', or 'casement'
                      retryable: false
                invalidStaadproPath:
                  summary: staadpro_path is not an existing file
                  value:
                    detail:
                      error_code: invalid_staadpro_path
                      message: "STAAD executable not found: C:\\missing\\Bentley.Staad.exe"
                      retryable: false
  /jobs/{job_id}:
    get:
      tags: [Jobs]
      operationId: getJobStatus
      summary: Get job status and progress
      description: |
        Poll this endpoint to drive a UI progress bar.

        - Continue while `status` is `queued` or `running`
        - Stop when `status` is `succeeded`, `failed`, or `canceled`
        - Call the result endpoint only when `result_ready` is `true`

        Non-terminal jobs cap `progress_percent` below 100. Terminal jobs always
        return 100.
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Current job status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobStatusResponse"
              examples:
                queued:
                  summary: Queued, not started
                  value:
                    job_id: 3f8c2a1e-9b4d-4e6a-8c1f-2a7b9d0e5f11
                    status: queued
                    progress_percent: 0
                    message: Job is not started yet.
                    attempts: 0
                    max_attempts: 3
                    created_at: "2026-08-22T17:00:00Z"
                    started_at: null
                    finished_at: null
                    result_ready: false
                    error: null
                    duration_seconds: null
                    working_directory: "C:\\STAADJobs\\3f8c2a1e-9b4d-4e6a-8c1f-2a7b9d0e5f11"
                    cleanup_performed: false
                running:
                  summary: Extraction in progress
                  value:
                    job_id: 3f8c2a1e-9b4d-4e6a-8c1f-2a7b9d0e5f11
                    status: running
                    progress_percent: 55
                    message: Running STAAD analysis.
                    attempts: 1
                    max_attempts: 3
                    created_at: "2026-08-22T17:00:00Z"
                    started_at: "2026-08-22T17:00:02Z"
                    finished_at: null
                    result_ready: false
                    error: null
                    duration_seconds: null
                    working_directory: "C:\\STAADJobs\\3f8c2a1e-9b4d-4e6a-8c1f-2a7b9d0e5f11"
                    cleanup_performed: false
                succeeded:
                  summary: Ready for result fetch
                  value:
                    job_id: 3f8c2a1e-9b4d-4e6a-8c1f-2a7b9d0e5f11
                    status: succeeded
                    progress_percent: 100
                    message: Extraction completed.
                    attempts: 1
                    max_attempts: 3
                    created_at: "2026-08-22T17:00:00Z"
                    started_at: "2026-08-22T17:00:02Z"
                    finished_at: "2026-08-22T17:01:15Z"
                    result_ready: true
                    error: null
                    duration_seconds: 73.2
                    working_directory: "C:\\STAADJobs\\3f8c2a1e-9b4d-4e6a-8c1f-2a7b9d0e5f11"
                    cleanup_performed: true
        "404":
          description: Unknown job id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetailEnvelope"
              example:
                detail:
                  error_code: job_not_found
                  message: "Job not found: 3f8c2a1e-9b4d-4e6a-8c1f-2a7b9d0e5f11"
                  retryable: false
  /jobs/{job_id}/abort:
    post:
      tags: [Jobs]
      operationId: abortJob
      summary: Abort a queued or running job
      description: |
        Cancels **queued** jobs immediately, or force-stops a **running** STAAD job
        by killing the extractor process tree and related STAAD processes so the
        worker can advance to the next queued request.

        Already-terminal jobs return their current status with an informational message.
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Abort accepted, or job already terminal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobAbortResponse"
              examples:
                canceledQueued:
                  summary: Queued job canceled
                  value:
                    job_id: 3f8c2a1e-9b4d-4e6a-8c1f-2a7b9d0e5f11
                    status: canceled
                    message: Job canceled before execution.
                canceledRunning:
                  summary: Running job force-stopped
                  value:
                    job_id: 3f8c2a1e-9b4d-4e6a-8c1f-2a7b9d0e5f11
                    status: canceled
                    message: Job canceled during execution.
                alreadySucceeded:
                  summary: Already finished
                  value:
                    job_id: 3f8c2a1e-9b4d-4e6a-8c1f-2a7b9d0e5f11
                    status: succeeded
                    message: Job already succeeded.
        "404":
          description: Unknown job id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetailEnvelope"
              example:
                detail:
                  error_code: job_not_found
                  message: "Job not found: 3f8c2a1e-9b4d-4e6a-8c1f-2a7b9d0e5f11"
                  retryable: false
...
components:
  parameters:
    JobId:
      name: job_id
      in: path
      required: true
      description: Job identifier returned by `POST /jobs`
      schema:
        type: string
        format: uuid
      example: 3f8c2a1e-9b4d-4e6a-8c1f-2a7b9d0e5f11
  schemas:
...
    JobStatus:
      type: string
      description: Job lifecycle state
      enum: [queued, running, succeeded, failed, canceled]
...
    ApiErrorCode:
      type: string
      description: Machine-readable error code
      examples:
        - job_not_found
        - file_too_large
        - invalid_input_file
        - invalid_timeout
        - invalid_max_retries
        - invalid_retry_delay
        - invalid_extraction_flow
        - invalid_staadpro_path
        - staad_attach_failed
        - job_failed
    ErrorBody:
      type: object
      description: Common application error object returned inside `detail`
      required: [error_code, message]
      properties:
        error_code:
          $ref: "#/components/schemas/ApiErrorCode"
        message:
          type: string
          description: Human-readable error summary
        retryable:
          type: boolean
          description: Whether the client may safely retry the same operation
          default: false
    ErrorDetailEnvelope:
      type: object
      description: FastAPI `HTTPException` wrapper used by most handler errors
      required: [detail]
      properties:
        detail:
          $ref: "#/components/schemas/ErrorBody"
    ValidationErrorItem:
      type: object
      description: Starlette / Pydantic validation error entry
      required: [loc, msg, type]
      properties:
        loc:
          type: array
          items:
            oneOf:
              - type: string
              - type: integer
          description: Location of the invalid field
        msg:
          type: string
        type:
          type: string
        input:
          description: Rejected input value (when available)
        ctx:
          type: object
          additionalProperties: true
    ValidationErrorResponse:
      type: object
      description: Framework validation failure (HTTP 422)
      required: [detail]
      properties:
        detail:
          type: array
          items:
            $ref: "#/components/schemas/ValidationErrorItem"
    ApiError:
      type: object
      description: |
        Structured failure attached to job status (and sanitized for API clients).
        `last_stderr_excerpt` is always stripped to `null` in HTTP responses.
      required: [error_code, message]
      properties:
        error_code:
          type: string
        message:
          type: string
        retryable:
          type: boolean
          default: false
        attempts:
          type: integer
          minimum: 0
          default: 0
        stage:
          type: string
          nullable: true
          description: Failure stage, e.g. `connection` or `execution`
        hint:
          type: string
          nullable: true
          description: Optional remediation guidance for operators
        details:
          type: object
          nullable: true
          additionalProperties: true
          description: |
            Sanitized subset of error metadata. Typical safe keys:
            `error_code`, `retryable`, `attempts`, `stage`, `hint`.
        last_stderr_excerpt:
          type: string
          nullable: true
          description: Always `null` in public API responses (stripped server-side)

    # ---------------------------------------------------------------------------
    # Health
    # ---------------------------------------------------------------------------
...
    JobCreateRequest:
      type: object
      required: [std_file]
      properties:
        std_file:
          type: string
          format: binary
          description: Required STAAD model upload. Filename must end with `.std`.
        staadpro_path:
          type: string
          nullable: true
          description: |
            Optional legacy absolute path to the STAAD executable. Accepted for
            backward compatibility; if provided, the path must exist as a file.
          example: "C:\\Program Files\\Bentley\\Engineering\\STAAD.Pro CONNECT Edition\\STAAD\\Bentley.Staad.exe"
        timeout_seconds:
          type: integer
          description: Per-attempt STAAD/extraction timeout in seconds. Must be `> 0`.
          default: 180
          exclusiveMinimum: 0
          example: 180
        max_retries:
          type: integer
          description: |
            Number of retries after the first attempt. Total attempts =
            `max_retries + 1`. Must be `>= 0`. Default from `STAAD_API_MAX_RETRIES`.
          default: 2
          minimum: 0
          example: 2
        retry_delay_seconds:
          type: integer
          description: |
            Delay between attempts in seconds. Must be `>= 1`. Default from
            `STAAD_API_RETRY_DELAY_SECONDS`.
          default: 20
          minimum: 1
          example: 20
        extraction_flow:
          type: string
          description: |
            Extraction mode. Prefer `standard`, `fully_unitized`, or `casement`.
            Aliases `unitized` / `fullyunitized` (and hyphen forms) are accepted for unitized.
            Hyphenated `casement` maps to `casement`.
          default: standard
          example: standard
        generation_request:
          type: string
          format: binary
          description: |
            Optional JSON file. For Fully Unitized: panel geometry and optional flow
            override. For Casement: optional flow override and `casement_profiles`
            member-id map. See `GenerationRequest`.
...
    JobCreateResponse:
      type: object
      required: [job_id, status, message]
      properties:
        job_id:
          type: string
          format: uuid
          description: Unique job identifier for status / result / abort calls
        status:
          allOf:
            - $ref: "#/components/schemas/JobStatus"
          description: Always `queued` on successful acceptance
          example: queued
        message:
          type: string
          default: Job accepted.

    # ---------------------------------------------------------------------------
    # Job status / abort
    # ---------------------------------------------------------------------------
    JobStatusResponse:
      type: object
      required:
        - job_id
        - status
        - progress_percent
        - message
        - attempts
        - max_attempts
        - created_at
        - result_ready
        - cleanup_performed
      properties:
        job_id:
          type: string
        status:
          $ref: "#/components/schemas/JobStatus"
        progress_percent:
          type: integer
          minimum: 0
          maximum: 100
          description: |
            UI progress value. Non-terminal jobs are capped below 100;
            terminal jobs always return 100.
        message:
          type: string
          description: User-facing status text for display next to a progress bar
        attempts:
          type: integer
          minimum: 0
          description: Completed / in-flight attempt count
        max_attempts:
          type: integer
          minimum: 0
          description: "`max_retries + 1` from job submission"
        created_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
          nullable: true
        finished_at:
          type: string
          format: date-time
          nullable: true
        result_ready:
          type: boolean
          description: "`true` only when `status` is `succeeded`"
        error:
          allOf:
            - $ref: "#/components/schemas/ApiError"
          nullable: true
          description: Present when the job failed (sanitized)
        duration_seconds:
          type: number
          format: float
          nullable: true
          description: Wall-clock duration once finished
        working_directory:
          type: string
          nullable: true
          description: Job working folder path on the API host
        cleanup_performed:
          type: boolean
          description: Whether post-job cleanup has run
    JobAbortResponse:
      type: object
      required: [job_id, status, message]
      properties:
        job_id:
          type: string
        status:
          $ref: "#/components/schemas/JobStatus"
        message:
          type: string
          description: Outcome message (`Job canceled before execution.` or `Job already {status}.`)

    # ---------------------------------------------------------------------------
    # Job result response variants
    # ---------------------------------------------------------------------------
...